+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.4.0
 - Events API
-- Locks 
+- Locks in storage 
+- request defaults
 
 ### 1.3.0
 - Notes for files